ai_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
scikit-learn==1.5.2
numpy==2.1.3
pydantic==2.9.2
